+++ packages/patrol/lib/src/platform/contracts/contracts.dart
@@ -290,17 +290,27 @@ class RunDartTestResponse with Equatable {
 
 @JsonSerializable()
 class ConfigureRequest with Equatable {
-  const ConfigureRequest({required this.findTimeoutMillis});
+  const ConfigureRequest({
+    required this.findTimeoutMillis,
+    this.androidDontSuppressAccessibilityServices,
+    this.androidRetrieveInteractiveWindows,
+  });
 
   factory ConfigureRequest.fromJson(Map<String, dynamic> json) =>
       _$ConfigureRequestFromJson(json);
 
   final int findTimeoutMillis;
+  final bool? androidDontSuppressAccessibilityServices;
+  final bool? androidRetrieveInteractiveWindows;
 
   Map<String, dynamic> toJson() => _$ConfigureRequestToJson(this);
 
   @override
-  List<Object?> get props => [findTimeoutMillis];
+  List<Object?> get props => [
+    findTimeoutMillis,
+    androidDontSuppressAccessibilityServices,
+    androidRetrieveInteractiveWindows,
+  ];
 }
 
 @JsonSerializable()