+++ packages/patrol/lib/src/platform/contracts/contracts.g.dart
@@ -69,10 +69,20 @@ const _$RunDartTestResponseResultEnumMap = {
 ConfigureRequest _$ConfigureRequestFromJson(Map<String, dynamic> json) =>
     ConfigureRequest(
       findTimeoutMillis: (json['findTimeoutMillis'] as num).toInt(),
+      androidDontSuppressAccessibilityServices:
+          json['androidDontSuppressAccessibilityServices'] as bool?,
+      androidRetrieveInteractiveWindows:
+          json['androidRetrieveInteractiveWindows'] as bool?,
     );
 
 Map<String, dynamic> _$ConfigureRequestToJson(ConfigureRequest instance) =>
-    <String, dynamic>{'findTimeoutMillis': instance.findTimeoutMillis};
+    <String, dynamic>{
+      'findTimeoutMillis': instance.findTimeoutMillis,
+      'androidDontSuppressAccessibilityServices':
+          instance.androidDontSuppressAccessibilityServices,
+      'androidRetrieveInteractiveWindows':
+          instance.androidRetrieveInteractiveWindows,
+    };
 
 OpenAppRequest _$OpenAppRequestFromJson(Map<String, dynamic> json) =>
     OpenAppRequest(appId: json['appId'] as String);